REAL is 32-bit on PostgreSQL, and it was silently truncating ISK
185 casts across 14 files read CAST(x AS REAL). On SQLite REAL is a
64-bit double. On PostgreSQL it is float4 -- roughly seven significant
digits -- so a wallet balance of 1,234,567,890,123 came back as about
1.2345679e12 and every figure derived from it was wrong.

Nothing failed. No error, no empty grid: the numbers were simply not the
numbers. That makes this worse than everything else found in this port,
all of which announced itself. It was found only by following an
aggregate type question into what the surrounding casts were doing.

DOUBLE PRECISION is the spelling both engines want: float8 on
PostgreSQL, and SQLite gives REAL affinity to any type name containing
DOUB, so it means there exactly what REAL did.

Aggregates return wider types on PostgreSQL too, and the rows reading
them did not expect it:

  COUNT(*) is bigint, not int, so the monthly summary could not read its
  own count. Cast to INTEGER at nine sites where the row field is an int.

  SUM over a bigint column is NUMERIC, not bigint, which is why the
  mining tab would not load. Cast to BIGINT.

And a timestamptz compared to an empty string in the project summary.
That guard made sense when the column was text on SQLite and an unset
date really was ''; on a server the column cannot hold '' and no
operator matches timestamptz <> text. IS NOT NULL already says it.

+++ Alarms/Conditions/MarketContractCondition.cs
@@ -215,7 +215,7 @@ public sealed class MarketContractCondition : IAlarmCondition
         // ordering on a number would put 9 above 10.
         await using var cmd = conn.CreateCommand();
         cmd.CommandText = $"""
-            SELECT c."ContractId", CAST(c."Price" AS REAL) AS total, i."Quantity",
+            SELECT c."ContractId", CAST(c."Price" AS DOUBLE PRECISION) AS total, i."Quantity",
                    c."DateExpired", c."Title", c."RegionId"
             FROM "EsiContracts" c
             JOIN "EsiContractItems" i ON i."ContractId" = c."ContractId"
@@ -223,13 +223,13 @@ public sealed class MarketContractCondition : IAlarmCondition
               AND c."Status" = 'outstanding'
               AND i."TypeId" = $type AND i."IsIncluded" = TRUE
               AND i."Quantity" >= $qty
-              AND CAST(c."Price" AS REAL) > 0
-              AND CAST(c."Price" AS REAL) / i."Quantity" <= $price
+              AND CAST(c."Price" AS DOUBLE PRECISION) > 0
+              AND CAST(c."Price" AS DOUBLE PRECISION) / i."Quantity" <= $price
               {(bundled ? "" : """
                 AND (SELECT COUNT(DISTINCT x."TypeId") FROM "EsiContractItems" x
                      WHERE x."ContractId" = c."ContractId" AND x."IsIncluded" = TRUE) = 1
                 """)}
-            ORDER BY CAST(c."Price" AS REAL) / i."Quantity"
+            ORDER BY CAST(c."Price" AS DOUBLE PRECISION) / i."Quantity"
             LIMIT {MaxOffers}
             """;
         cmd.AddWithValue("$type", typeId);
